Reorganize preference pages
The new layout contains only the pages Compiler, Debug, Editor and
Logging. All other remaining pages are grouped under the Editor page.

Furthermore, a new consistent name scheme for the preference page ids
is applied. The scheme is as follows:

    org.scalaide.ui.preferences.<rootlevel-page>.<sublevel-page>

SyntaxColouringPreferencePage is renamed to SyntaxColoringPreferencePage
