+++ org.scala-ide.sdt.core/src/org/scalaide/logging/LogManager.scala
@@ -1,17 +1,12 @@
 package org.scalaide.logging
 
-import org.scalaide.logging.log4j.Log4JFacade
-import org.scalaide.core.ScalaPlugin
-import org.eclipse.jdt.internal.ui.viewsupport.IProblemChangedListener
-import org.eclipse.jface.util.IPropertyChangeListener
 import org.eclipse.jface.util.PropertyChangeEvent
-import org.eclipse.jdt.ui.PreferenceConstants
+import org.scalaide.core.ScalaPlugin
+import org.scalaide.logging.log4j.Log4JFacade
+import org.scalaide.logging.ui.preferences.LoggingPreferenceConstants._
 import org.scalaide.util.internal.eclipse.SWTUtils
-import java.io.File
-import org.eclipse.core.resources.ResourcesPlugin
 
 object LogManager extends Log4JFacade with HasLogger {
-  import ui.properties.LoggingPreferenceConstants._
 
   private def updateLogLevel(event: PropertyChangeEvent): Unit = {
     if (event.getProperty == LogLevel) {
